+++ Fruitmarket.Domain/Entities/OrderItem.cs
@@ -10,5 +10,6 @@ public sealed class OrderItem : BaseEntity
     public Product? Product { get; set; }
     public string ProductName { get; set; } = string.Empty;
     public decimal UnitPrice { get; set; }
-    public int Quantity { get; set; }
+    // Decimal to preserve the exact gram-based quantity ordered (0.25 = 250g).
+    public decimal Quantity { get; set; }
 }
